+++ src/main/kotlin/icu/windea/pls/lang/editor/ParadoxGameTypeNotDeclaredInMetadataJsonEditorNotificationProvider.kt
@@ -23,6 +23,7 @@ import javax.swing.JComponent
  * 如果游戏类型未在 `.metadata/metadata.json` 中声明，则为模组文件提供通知，以便快速配置（VIC3 / EU5）。
  * 仅适用于项目中的文本文件。
  */
+@Suppress("UnstableApiUsage")
 class ParadoxGameTypeNotDeclaredInMetadataJsonEditorNotificationProvider : EditorNotificationProvider {
     override fun collectNotificationData(project: Project, file: VirtualFile): Function<in FileEditor, out JComponent?>? {
         val fileInfo = file.fileInfo ?: return null