+++ src/main/kotlin/icu/windea/pls/model/tools/ParadoxModInfo.kt
@@ -1,7 +1,7 @@
 package icu.windea.pls.model.tools
 
 import icu.windea.pls.core.orNull
-import icu.windea.pls.lang.analysis.ParadoxMetadataManager
+import icu.windea.pls.lang.analysis.ParadoxMetadataUtil
 import icu.windea.pls.lang.settings.ParadoxModDependencySettingsState
 import icu.windea.pls.lang.settings.ParadoxModDescriptorSettingsState
 import icu.windea.pls.model.ParadoxModSource
@@ -45,7 +45,7 @@ fun ParadoxModDependencySettingsState.toModInfo(): ParadoxModInfo? {
 
 fun ParadoxModInfo.toModDependency(): ParadoxModDependencySettingsState? {
     val modDirectory = modDirectory?.orNull() ?: return null
-    val modInfo = ParadoxMetadataManager.getModInfoFromModDirectory(modDirectory) // 确保模组信息被解析
+    val modInfo = ParadoxMetadataUtil.getModInfoFromModDirectory(modDirectory) // 确保模组信息被解析
     if (modInfo == null) return null
     return ParadoxModDependencySettingsState().also {
         it.modDirectory = modDirectory